+++ src/ScottPlot5/ScottPlot5/Stylers/PlotStyler.cs
@@ -1,12 +1,8 @@
 ﻿namespace ScottPlot.Stylers;
 
-/// <summary>
-/// Helper methods for applying common styling options to plots
-/// </summary>
-public class PlotStyler(Plot plot)
+[Obsolete("This class is deprecated. See the ScottPlot Cookbook for styling information.", true)]
+public class PlotStyler()
 {
-    private readonly Plot Plot = plot;
-
     [Obsolete("This method is deprecated. Assign Plot.FigureBackground.Color instead.", true)]
     public void Background(Color figure, Color data) { }
 
@@ -25,21 +21,6 @@ public class PlotStyler(Plot plot)
     [Obsolete("This method is deprecated. Call Plot.Font.Automatic() instead.", true)]
     public void SetBestFonts() { }
 
-    /// <summary>
-    /// Reset colors and palette do a dark mode style
-    /// </summary>
-    public void DarkMode()
-    {
-        Plot.Add.Palette = new Palettes.Penumbra();
-
-        Plot.Axes.Color(Color.FromHex("#d7d7d7"));
-
-        Plot.Grid.MajorLineColor = Color.FromHex("#404040");
-        Plot.FigureBackground.Color = Color.FromHex("#181818");
-        Plot.DataBackground.Color = Color.FromHex("#1f1f1f");
-
-        Plot.Legend.BackgroundColor = Color.FromHex("#404040");
-        Plot.Legend.FontColor = Color.FromHex("#d7d7d7");
-        Plot.Legend.OutlineColor = Color.FromHex("#d7d7d7");
-    }
+    [Obsolete("This method is deprecated. Call Plot.SetStyle() instead.", true)]
+    public void DarkMode() { }
 }